Change compare operator in PreferenceFactory to functon equals(String). Some changes weres done to match the simultador with the algortim and... Changes in class GA_main.java to run the tests.

+++ src/jmetal/interactive/core/Preference.java
@@ -28,7 +28,7 @@ public abstract class Preference {
 	 * @return
 	 * @throws JMException 
 	 */
-	public abstract int evaluate(Solution solution) throws JMException;
+	public abstract double evaluate(Solution solution) throws JMException;
 	/**
 	 * 
 	 * @param weight
