+++ src/jmetal/interactive/preferences/PositioningAfter.java
@@ -25,14 +25,14 @@ public class PositioningAfter extends Preference{
 	}
 
 	@Override
-	public int evaluate(Solution solution) throws JMException {
+	public double evaluate(Solution solution) throws JMException {
 		Variable [] variables = solution.getDecisionVariables();
 		if(variables[r1].getValue() - k >= distance
 				|| (variables[r1].getValue() == 0)) {
-			return 1;
+			return 1.0;
 		}
 			
-		return 0;
+		return 0.0;
 	}
 	
 	@Override
