+++ src/jmetal/interactive/preferences/PositioningFollow.java
@@ -26,14 +26,14 @@ public class PositioningFollow extends Preference{
 	}
 
 	@Override
-	public int evaluate(Solution solution) throws JMException {
+	public double evaluate(Solution solution) throws JMException {
 		Variable [] variables = solution.getDecisionVariables();
 		if((variables[r1].getValue() - variables[r2].getValue() >= distance) 
 				|| (variables[r1].getValue() == 0 && variables[r2].getValue() > 0)) {
-			return 1;
+			return 1.0;
 		}
 			
-		return 0;
+		return 0.0;
 	}
 	
 	@Override
