+++ src/jmetal/interactive/preferences/PositioningIn.java
@@ -23,14 +23,14 @@ public class PositioningIn extends Preference{
 	}
 
 	@Override
-	public int evaluate(Solution solution) throws JMException {
+	public double evaluate(Solution solution) throws JMException {
 		Variable [] variables = solution.getDecisionVariables();
 		
 		if(variables[r1].getValue() == k) {
-			return 1;
+			return 1.0;
 		}
 			
-		return 0;
+		return 0.0;
 	}
 	
 	@Override
